refactor(play): rename PlayController to PlayProvider and update related imports
