+++ lib/pages/play/providers/video_source_provider.dart
@@ -7,8 +7,8 @@ import 'package:anime_flow/crawler/itme/crawler_config_item.dart';
 import 'package:anime_flow/features/search_result_rank_service.dart';
 import 'package:anime_flow/models/play/video/episode_resources_item.dart';
 import 'package:anime_flow/models/play/video/resources_item.dart';
-import 'package:anime_flow/pages/play/controller/play_controller.dart';
-import 'package:anime_flow/pages/play/provider/episodes_provider.dart';
+import 'package:anime_flow/pages/play/providers/play_provider.dart';
+import 'package:anime_flow/pages/play/providers/episodes_provider.dart';
 import 'package:anime_flow/routes/provider/routes_args.dart';
 import 'package:anime_flow/repository/play_repository.dart';
 import 'package:anime_flow/providers/video/providers.dart';
@@ -16,7 +16,7 @@ import 'package:anime_flow/utils/crawl_config.dart';
 import 'package:anime_flow/utils/logger.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-part 'video_source_controller.g.dart';
+part 'video_source_provider.g.dart';
 
 class VideoSourceState {
   const VideoSourceState({
