+++ lib/pages/play/video/gesture/mobile_gesture_detector.dart
@@ -1,7 +1,7 @@
 import 'package:anime_flow/constants/storage_key.dart';
 import 'package:anime_flow/models/enums/video_controls_icon_type.dart';
-import 'package:anime_flow/pages/play/controller/play_controller.dart';
-import 'package:anime_flow/pages/play/controller/video_ui_controller.dart';
+import 'package:anime_flow/pages/play/providers/play_provider.dart';
+import 'package:anime_flow/pages/play/providers/video_ui_provider.dart';
 import 'package:anime_flow/repository/storage.dart';
 import 'package:anime_flow/utils/vibrate.dart';
 import 'package:flutter/material.dart';
