+++ lib/pages/play/video/ui/video_ui_components.dart
@@ -1,5 +1,5 @@
-import 'package:anime_flow/pages/play/controller/play_controller.dart';
-import 'package:anime_flow/pages/play/controller/video_ui_controller.dart';
+import 'package:anime_flow/pages/play/provider/play_provider.dart';
+import 'package:anime_flow/pages/play/provider/video_ui_provider.dart';
 import 'package:anime_flow/utils/format_time_util.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
